backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
sqlalchemy
sqlmodel
alembic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
faiss-cpu
sentence-transformers
openai
python-dotenv
pytest
pip-audit
